+++ util/lev_main.c
@@ -380,9 +380,9 @@ void
 yyerror(const char *s)
 {
     char *e = ((char *)s + strlen(s) - 1);
-    (void) fprintf(stderr, "%s: line %d, pos %zu : %s",
+    (void) fprintf(stderr, "%s: line %d, pos %lu : %s",
                    fname, line_number,
-                   token_start_pos-strlen(curr_token), s);
+                   (unsigned long)(token_start_pos-strlen(curr_token)), s);
     if (*e != '.' && *e != '!')
         (void) fprintf(stderr, " at \"%s\"", curr_token);
     (void) fprintf(stderr, "\n");
